pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "arxiv-research-demo"
version = "0.1.0"
description = "Demo combining ArXiv search, extraction, and GitHub gist creation"
authors = [
    {name = "Research Demo", email = "demo@example.com"},
]
requires-python = ">=3.10"
dependencies = [
    "arxiv>=2.1.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "fastmcp>=2.10.0",
    "mirascope[openai]>=1.25.4",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
]
